requirements.txt
nltk
multidict
bs4
wordcloud
matplotlib
Pillow